All repository. Schema created from logic.

+++ xpacs/src/main/java/org/cardiacatlas/xpacs/service/PatientInfoRepository.java
@@ -3,6 +3,6 @@ package org.cardiacatlas.xpacs.service;
 import org.cardiacatlas.xpacs.domain.PatientInfo;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface PatientInfoRepository extends PagingAndSortingRepository<PatientInfo, Integer> {
+public interface PatientInfoRepository extends PagingAndSortingRepository<PatientInfo, Long> {
 
 }
